Extract filter and sorting
