+++ src/popup/interface.js
@@ -1,9 +1,9 @@
 module.exports = Interface;
 
 const m = require("mithril");
-const FuzzySort = require("fuzzysort");
 const Moment = require("moment");
 const SearchInterface = require("./searchinterface");
+const helpers = require("../helpers");
 
 const LATEST_NATIVE_APP_VERSION = 3000003;
 
@@ -144,188 +144,11 @@ function view(ctl, params) {
 /**
  * Run a search
  *
- * @param string s Search string
+ * @param string searchQuery Search query
  * @return void
  */
-function search(s) {
-    var self = this;
-    var fuzzyFirstWord = s.substr(0, 1) !== " ";
-    s = s.trim();
-
-    // get candidate list
-    var candidates = this.logins.map(candidate => {
-        let lastSlashIndex = candidate.login.lastIndexOf("/") + 1;
-        return Object.assign(candidate, {
-            path: candidate.login.substr(0, lastSlashIndex),
-            display: candidate.login.substr(lastSlashIndex)
-        });
-    });
-    var mostRecent = null;
-    if (this.currentDomainOnly) {
-        var recent = candidates.filter(function(login) {
-            if (login.recent.count > 0) {
-                // find most recently used login
-                if (!mostRecent || login.recent.when > mostRecent.recent.when) {
-                    mostRecent = login;
-                }
-                return true;
-            }
-            return false;
-        });
-        var remainingInCurrentDomain = candidates.filter(
-            login => login.inCurrentDomain && !login.recent.count
-        );
-        candidates = recent.concat(remainingInCurrentDomain);
-    }
-    candidates.sort((a, b) => {
-        // show most recent first
-        if (a === mostRecent) {
-            return -1;
-        }
-        if (b === mostRecent) {
-            return 1;
-        }
-
-        // sort by frequency
-        var countDiff = b.recent.count - a.recent.count;
-        if (countDiff) {
-            return countDiff;
-        }
-
-        // sort by specificity, only if filtering for one domain
-        if (this.currentDomainOnly) {
-            var domainLevelsDiff =
-                (b.login.match(/\./g) || []).length - (a.login.match(/\./g) || []).length;
-            if (domainLevelsDiff) {
-                return domainLevelsDiff;
-            }
-        }
-
-        // sort alphabetically
-        return a.login.localeCompare(b.login);
-    });
-
-    if (s.length) {
-        let filter = s.split(/\s+/);
-        let fuzzyFilter = fuzzyFirstWord ? filter[0] : "";
-        let substringFilters = filter.slice(fuzzyFirstWord ? 1 : 0).map(w => w.toLowerCase());
-
-        // First reduce the list by running the substring search
-        substringFilters.forEach(function(word) {
-            candidates = candidates.filter(c => c.login.toLowerCase().indexOf(word) >= 0);
-        });
-
-        // Then run the fuzzy filter
-        let fuzzyResults = {};
-        if (fuzzyFilter) {
-            candidates = FuzzySort.go(fuzzyFilter, candidates, {
-                keys: ["login", "store.name"],
-                allowTypo: false
-            }).map(result => {
-                fuzzyResults[result.obj.login] = result;
-                return result.obj;
-            });
-        }
-
-        // Finally highlight all matches
-        candidates = candidates.map(c => highlightMatches(c, fuzzyResults, substringFilters));
-    }
-
-    // Prefix root entries with slash to let them have some visible path
-    candidates.forEach(c => {
-        c.path = c.path || "/";
-    });
-
-    this.results = candidates;
-}
-
-/**
- * Highlight filter matches
- *
- * @since 3.0.0
- *
- * @param object entry password entry
- * @param object fuzzyResults positions of fuzzy filter matches
- * @param array substringFilters list of substring filters applied
- * @return object entry with highlighted matches
- */
-function highlightMatches(entry, fuzzyResults, substringFilters) {
-    // Add all positions of the fuzzy search to the array
-    let matches = (fuzzyResults[entry.login] && fuzzyResults[entry.login][0]
-        ? fuzzyResults[entry.login][0].indexes
-        : []
-    ).slice();
-
-    // Add all positions of substring searches to the array
-    let login = entry.login.toLowerCase();
-    for (let word of substringFilters) {
-        let startIndex = login.indexOf(word);
-        for (let i = 0; i < word.length; i++) {
-            matches.push(startIndex + i);
-        }
-    }
-
-    // Prepare the final array of matches before
-    matches = sortUnique(matches, (a, b) => a - b);
-
-    const OPEN = "<em>";
-    const CLOSE = "</em>";
-    let highlighted = "";
-    var matchesIndex = 0;
-    var opened = false;
-    for (var i = 0; i < entry.login.length; ++i) {
-        var char = entry.login[i];
-
-        if (i == entry.path.length) {
-            if (opened) {
-                highlighted += CLOSE;
-            }
-            var path = highlighted;
-            highlighted = "";
-            if (opened) {
-                highlighted += OPEN;
-            }
-        }
-
-        if (matches[matchesIndex] === i) {
-            matchesIndex++;
-            if (!opened) {
-                opened = true;
-                highlighted += OPEN;
-            }
-        } else {
-            if (opened) {
-                opened = false;
-                highlighted += CLOSE;
-            }
-        }
-        highlighted += char;
-    }
-    if (opened) {
-        opened = false;
-        highlighted += CLOSE;
-    }
-    let display = highlighted;
-
-    return Object.assign(entry, {
-        path: path,
-        display: display
-    });
-}
-
-/**
- * Sort and remove duplicates
- *
- * @since 3.0.0
- *
- * @param array array items to sort
- * @param function comparator sort comparator
- * @return array sorted items without duplicates
- */
-function sortUnique(array, comparator) {
-    return array
-        .sort(comparator)
-        .filter((elem, index, arr) => index == !arr.length || arr[index - 1] != elem);
+function search(searchQuery) {
+    this.results = helpers.filterSortLogins(this.logins, searchQuery, this.currentDomainOnly);
 }
 
 /**
